Move code to reduce size of cli.py.
